+++ msr/graph/graph_lib/__init__.py
@@ -1,1 +1,2 @@
 from .graph_lib import * # I know this is bad practice, but it's so convenient
+from .generate import generate_all_graphs_on_n_vertices