+++ msr/graph/graph_lib/graph_lib.py
@@ -1,5 +1,7 @@
 from ..simple_undirected_graph import simple_undirected_graph
 
+### variable number of vertices ###############################################
+
 def empty(num_vert: int) -> simple_undirected_graph:
 	return simple_undirected_graph(num_vert)
 
@@ -34,9 +36,19 @@ def star(num_vert: int) -> simple_undirected_graph:
 		G.add_edge(i, num_vert - 1)
 	return G
 
+### fixed number of vertices ##################################################
+
 def house() -> simple_undirected_graph:
 	G = cycle(num_vert=5)
 	G.add_edge(1, 4)
+	return G
+
+def petersen() -> simple_undirected_graph:
+	G = empty(num_vert=10)
+	for i in range(5):
+		G.add_edge(i, (i + 1) % 5)
+		G.add_edge(i + 5, (i + 2) % 5 + 5)
+		G.add_edge(i, i + 5)
 	return G
 
 def triforce() -> simple_undirected_graph: